[refactor] : 작성한 노트가 없어도 노트 추가 버튼이 보일 수 있도록 위치 변경
